Messy, but works???

+++ rmirro.py
@@ -6,7 +6,10 @@ import shutil
 import json
 import urllib.request
 import time
+import datetime
 import uuid
+
+DRY_RUN = False
 
 def pc_run(cmd):
     output = subprocess.getoutput(cmd)
@@ -15,12 +18,11 @@ def pc_run(cmd):
 def rm_run(cmd):
     return pc_run(f"ssh {RM_SSH_NAME} {cmd}")
 
-def ssh_ip():
-    return pc_run("ssh remarkable -v exit 2>&1 | grep 'Connecting to' | cut -d' ' -f4") # e.g. 10.11.99.1
-
 RM_SSH_NAME = "remarkable"
-RM_SSH_IP = ssh_ip()
+RM_SSH_IP = pc_run(f"ssh {RM_SSH_NAME} -v exit 2>&1 | grep 'Connecting to' | cut -d' ' -f4") # e.g. 10.11.99.1
 RM_CONTENT_PATH = "/home/root/.local/share/remarkable/xochitl"
+
+print(f"Connected to {RM_SSH_NAME} at {RM_SSH_IP}")
 
 PC_LAST_SYNC_PATH = RM_SSH_NAME + "/.last_sync"
 
@@ -49,26 +51,43 @@ def rmnewer(file):
     pc_path = RM_SSH_NAME + rm_path
     pc_exists = os.path.exists(pc_path)
     pc_mtime = int(os.path.getmtime(pc_path) * 1000) if pc_exists else 0 # ms
+    #print((rm_mtime - pc_mtime) / 1000 / 3600 / 24)
     return rm_mtime > pc_mtime
 
 def download_file(file, pcpath):
     if file["type"] == "CollectionType":
-        print("MDIR", file["path"])
-        pc_run(f"mkdir {pcpath}")
+        if os.path.exists(pcpath):
+            print("SKIP", file["path"])
+        else:
+            print("MDIR", file["path"])
+            if DRY_RUN:
+                return
+            pc_run(f"mkdir \"{pcpath}\"")
+            mtime = int(file["lastModified"]) / 1000 # s
+            os.utime(pcpath, (mtime, mtime)) # sync with access/modification times from RM
     elif file["type"] == "DocumentType":
         if rmnewer(file):
             # download
             print("PULL", file["path"])
-            url = f"http://{RM_SSH_IP}/download/{file['id']}/placeholder"
-            urllib.request.urlretrieve(url, filename=pcpath)
-            atime = int(file["lastOpened"]) / 1000 # s
-            mtime = int(file["lastModified"]) / 1000 # s
-            os.utime(pcpath, (atime, mtime)) # sync with access/modification times from RM
+            if DRY_RUN:
+                return
+            try:
+                url = f"http://{RM_SSH_IP}/download/{file['id']}/placeholder"
+                urllib.request.urlretrieve(url, filename=pcpath)
+                atime = int(file["lastOpened"]) / 1000 # s
+                mtime = int(file["lastModified"]) / 1000 # s
+                os.utime(pcpath, (atime, mtime)) # sync with access/modification times from RM
+            except Exception as e:
+                print(f"ERROR Could not download {url} from {RM_SSH_NAME}.")
+                print(f"      Enable \"Settings > Storage > USB web interface\" on {RM_SSH_NAME}!")
+                exit()
         else:
             # copy cached
             print("SKIP", file["path"])
-            pcpath_cached = RM_SSH_NAME + rmpath
-            shutil.copy2(pcpath_cached, pcpath)
+            if DRY_RUN:
+                return
+            #pcpath_cached = RM_SSH_NAME + rmpath
+            #shutil.copy2(pcpath_cached, pcpath)
     else:
         raise f"Unknown file type: {file['type']}"
 
@@ -99,10 +118,32 @@ def download_files(skiproot=True, skipfolders=False, skipolder=False):
         
         files[id] = file
 
+    """
+    # If a folder is trashed on RM, then its contents are NOT marked as trash
+    # Thus, "redefine" a trashed file as a file that is in the trash or that has a parent in the trash
+    def is_in_trash(fileid):
+        print(files[fileid])
+        if fileid == "": # root directory, is not in trash
+            return False
+        elif fileid
+        elif fileid == "trash":
+            return True
+        else:
+            return is_in_trash(files[fileid]["parent"])
+
+    # Remove files whose parents
+    for fileid in files:
+        if is_in_trash(fileid):
+            print("delete", files[fileid])
+            del files[fileid]
+    """
+
     # Build file tree edges
-    for fileid, file in files.items():
-        skip = fileid == ""
-        if skip:
+    for fileid, file in list(files.items()): # list copies keys and values, allowing deletion of dictionary entries in the loop
+        if fileid == "":
+            continue # skip root
+        if file["parent"] not in files:
+            del files[fileid] # e.g. forget files that are in a deleted folder
             continue
         parentname = file["parent"]
         parent = files[parentname]
@@ -154,6 +195,7 @@ def download_files(skiproot=True, skipfolders=False, skipolder=False):
         rm_files_by_pc_path[pc_path] = rm_file
 
     print("Iterating over PC files:")
+    uploaded_files = False
     for pc_path in pc_paths(skip_root=True, skip_hidden=True, top_down=False):
         is_dir = is_directory(pc_path)
         is_root = pc_path == RM_SSH_NAME
@@ -163,38 +205,41 @@ def download_files(skiproot=True, skipfolders=False, skipolder=False):
         pc_parent_path = None if is_root else os.path.dirname(pc_path)
         if pc_parent_path == RM_SSH_NAME:
             pc_parent_path += "/" # TODO: remove special case
-        rm_parent_file = None if is_root else rm_files_by_pc_path[pc_parent_path]
+        rm_parent_file = None if is_root or pc_parent_path not in rm_files_by_pc_path else rm_files_by_pc_path[pc_parent_path]
         rm_mtime = 0 if not rm_exists else int(rm_file["lastModified"])
         pc_mtime = int(os.path.getmtime(pc_path) * 1000)
         if rm_exists and pc_mtime > rm_mtime and not is_dir:
-            rm_file = upload_file(pc_path, rm_file, rm_parent_file) #
+            rm_file = upload_file(pc_path, rm_file, rm_parent_file) # update
+            uploaded_files = True
         elif not rm_exists:
+            if not is_dir and os.path.getctime(pc_path) > os.path.getmtime(pc_path):
+                os.utime(pc_path, (os.path.getatime(pc_path), os.path.getctime(pc_path))) # set mtime = ctime
+            pc_mtime = int(os.path.getmtime(pc_path) * 1000)
             # TODO: check prev sync time, then delete or upload
-            diff = pc_mtime - last_sync
-            print(diff / 1000, "sec")
-            if pc_mtime > last_sync:
-                rm_file = upload_file(pc_path, rm_file, rm_parent_file) #
+            print((pc_mtime - last_sync) / 1000, "sec")
+            print("PC mtime ", datetime.datetime.fromtimestamp(pc_mtime / 1000))
+            print("Last sync", datetime.datetime.fromtimestamp(last_sync / 1000))
+            if pc_mtime > last_sync and rm_parent_file: # delete if file does not have a parent
+                rm_file = upload_file(pc_path, rm_file, rm_parent_file) # create
+                uploaded_files = True
             else:
-                delete_file_pc(pc_path)
+                delete_file_pc(pc_path, pc_parent_path)
         rm_files_by_pc_path[pc_path] = rm_file
-        # TODO: add file to rm_files_by_pc_path
 
-    #print("Restarting RM interface")
-    #rm_run("systemctl restart xochitl")
+    if uploaded_files:
+        # Remarkable system needs to be reloaded to show uploaded files
+        print("Restarting RM interface")
+        rm_run("systemctl restart xochitl")
 
-    write_last_sync_time()
+    if not DRY_RUN:
+        write_last_sync_time()
 
     print("Complete")
 
-def write_file(path, text, transfer=False):
-    with open("/tmp/rmirro/" + path, "w") as file:
+def write_file(path, text):
+    with open(f"/tmp/rmirro-{path}", "w") as file:
         file.write(text)
-    print("wrote", path)
-
-    if transfer:
-        # TODO: reuse transfered metadata files?
-        pc_run("scp /tmp/rmirro/{path} {RM_SSH_NAME}:{RM_CONTENT_PATH}/{path}")
-        print("transfered", path)
+    pc_run(f"scp /tmp/rmirro-{path} {RM_SSH_NAME}:{RM_CONTENT_PATH}/{path}")
 
 def write_content(id):
     pass # TODO: needed?
@@ -206,77 +251,30 @@ def read_metadata(id):
     mdfile = open(f"{RM_SSH_NAME}_metadata/{id}.metadata", "r")
     return json.load(mdfile)
 
-# Create file (including directories)
-def create_file(path, parent):
-    curtime = int(time.time() * 1000) # ms
-    filename = os.path.basename(path) # e.g. "document.pdf"
-    name, ext = os.path.splitext(filename) # e.g. ("document", ".pdf")
-    print(name, ext)
-
-    metadata = {
-        "visibleName": name,
-        "parent": parent["id"],
-        "lastModified": str(curtime),
-        "modified": False,
-        "metadatamodified": False,
-        "deleted": False,
-        "pinned": False,
-        "synced": False,
-        "version": 0,
-    }
-    if ext == "": # directory
-        metadata["type"] = "CollectionType"
-    elif ext == ".pdf":
-        metadata["type"] = "DocumentType"
-        metadata["lastOpened"] = metadata["lastModified"]
-    else:
-        raise "Unknown filetype: " + ext
-
-    #content = {"tags": []}
-    id = uuid.uuid4()
-
-    # TODO: generate id and write file
-    # TODO: move files to RM
-    #write_metadata(id, metadata)
-    #write_file(f"{id}.metadata", json.dumps(metadata, indent=4) + "\n")
-    #write_file(f"{id}.content", json.dumps(content, indent=4).replace("[]", "[\n    ]") + "\n")
-    # TODO: must also return file and add to my tree
-    return metadata
-
-def update_file(file):
-    # TODO: upload file, and update metadata
-    # TODO: update metadata
-
-    # store needed "extra" metadata
-    id = file["id"]
-    mtime = int(os.path.getmtime(RM_SSH_NAME + file["path"]) * 1000) # ms
-
-    # re-read file with original metadata
-    file = read_metadata(id)
-    
-    # modify relevant metadata
-    # TODO: does other metadata or other files need to be modified?
-    file["lastModified"] = str(mtime)
-    file["lastOpened"] = str(mtime)
-    file["synced"] = False
-
-    #write_metadata(id, metadata)
-
-    # TODO: move to RM
-    # TODO: also move the actual file
-
 def is_directory(path):
     return path[-4:] not in [".pdf"]
 
-def delete_file_pc(pc_path):
+def delete_file_pc(pc_path, pc_parent_path):
     print("RMPC", pc_path)
+    if DRY_RUN:
+        return
+
+    atime = os.path.getatime(pc_parent_path)
+    mtime = os.path.getmtime(pc_parent_path)
+
     if is_directory(pc_path):
         os.rmdir(pc_path)
     else:
         os.remove(pc_path)
+        
+    os.utime(pc_parent_path, (atime, mtime)) # restore old (atime, mtime), since we use mtime to determine how to sync files/directories
 
 def upload_file(pc_path, rm_file, rm_parent_file):
     print("PUSH", pc_path)
+    if DRY_RUN:
+        return
+
+    # TODO: NEVER overwrite a note on Remarkable with a PDF
 
     filename = os.path.basename(pc_path) # e.g. "document.pdf"
     name, ext = os.path.splitext(filename) # e.g. ("document", ".pdf")
@@ -301,45 +299,20 @@ def upload_file(pc_path, rm_file, rm_parent_file):
             rm_file["lastOpened"] = str(pc_mtime)
     else:
         id = rm_file["id"]
-        rm_file = read_metadata(id)
+        rm_file = read_metadata(id) # read original metadata, don't pollute with our own
 
     rm_file["lastModified"] = str(pc_mtime)
     rm_file["synced"] = False
 
-    # TODO: write first,
+    # send file
+    write_file(f"{id}.metadata", json.dumps(rm_file, indent=4) + "\n")
+    write_file(f"{id}.content", json.dumps({}, indent=4) + "\n")
+    if rm_file["type"] == "DocumentType":
+        print(f"scp \"{pc_path}\" {RM_SSH_NAME}:{RM_CONTENT_PATH}/{id}.pdf") # send file
+        pc_run(f"scp \"{pc_path}\" {RM_SSH_NAME}:{RM_CONTENT_PATH}/{id}.pdf") # send file
 
-    # TODO: add id, AFTER writing file
-    rm_file["id"] = id
+    rm_file["id"] = id # add id, AFTER writing file
     return rm_file
-
-def upload_files(files):
-    files_by_path = {file["path"]: file for id, file in files.items()}
-
-    for dirpath, dirs, files in os.walk(f"{RM_SSH_NAME}"):
-        rm_path = dirpath[len(f"{RM_SSH_NAME}"):]
-        rm_parent_id = "" if dirpath == "/" else files_by_path
-        print(dirpath)
-        if rm_path != "" and not rm_path in files_by_path:
-            # directory does not exist on remarkable
-            # TODO: create t
-            #create_directory(0)
-            print(f"CREATE DIR: {rm_path}")
-
-        for file in files:
-            filepath = dirpath + "/" + file
-            rm_path = filepath[len(f"{RM_SSH_NAME}"):]
-            if rm_path in files_by_path:
-                # file already exists on remarkable
-                # TODO: update file if outdated
-                print(f"UPDATE IF OUTDATED: {rm_path}")
-                pass
-            else:
-                # file does not exist on remarkable
-                # TODO: create file
-                print(f"CREATE FILE: {rm_path}")
-                pass
-
-    # TODO: restart xochitl thing
 
 if __name__ == "__main__":
     #create_file("/a/B/c.pdf", {"id": "cjfdks"})
